update.
PAY ATTENTION TO THIS:

Add a database migration that includes the following: a ‘Company’ table to CoreData that has attribute name:String. The Company table should have a one-to-many relationship with the existing Employee table. Use the name of the company to show a welcome message above the “Start” button.

Also,

•	Organize your code into separate layers/packages/folders to demonstrate a clear separation of network calls/business logic from the UI.
	•	The api call mocking could be as simple as a method call. As long as the method lives in a separate layer than the UI layer and the call itself happens on a background thread.
	•	Handle background work in any way you prefer.
	•	DO NOT use storyboards.
	•	Design the UI with full-screen display, paying attention to proper padding and positioning.
	•	Don’t spend a lot of time on aesthetics such as colors, icons, themes, and animations.
	•	Use CoreData. 